refactor(pi-coding-agent): migrate from CLI subprocess to SDK
Replace the Pi CLI subprocess spawning with direct calls to
@mariozechner/pi-coding-agent SDK's createAgentSession API.
Remove the deprecated pi-agent-sdk provider entirely.

Key changes:
- pi-coding-agent provider: uses SDK directly (createAgentSession + session.subscribe)
- pi-cli provider: new subprocess provider for users who prefer the CLI binary
- @mariozechner/pi-coding-agent is now an optional peer dependency with auto-install
  prompt on first use (interactive TTY only)
- Remove pi-agent-sdk provider, config types, and registry entries
- Drop executable field from pi-coding-agent (not needed with SDK)
- Map config apiKey to provider-specific env vars for SDK consumption

Closes #704
