+++ .agentv/targets.yaml
@@ -4,7 +4,6 @@
 targets:
   - name: default
     provider: pi-coding-agent
-    executable: ${{ PI_CLI_PATH }}
     subprovider: openrouter
     model: z-ai/glm-4.7
     api_key: ${{ OPENROUTER_API_KEY }}
