+++ packages/core/package.json
@@ -48,8 +48,6 @@
     "@ai-sdk/openai": "^3.0.0",
     "@anthropic-ai/claude-agent-sdk": "^0.2.49",
     "@github/copilot-sdk": "^0.1.25",
-    "@mariozechner/pi-agent-core": "^0.54.2",
-    "@mariozechner/pi-ai": "^0.54.2",
     "@openai/codex-sdk": "^0.104.0",
     "@openrouter/ai-sdk-provider": "^2.3.1",
     "ai": "^6.0.0",
@@ -66,6 +64,12 @@
     "@opentelemetry/resources": "^2.5.1",
     "@opentelemetry/sdk-trace-node": "^2.5.1",
     "@opentelemetry/semantic-conventions": "^1.39.0"
+  },
+  "peerDependencies": {
+    "@mariozechner/pi-coding-agent": "^0.62.0"
+  },
+  "peerDependenciesMeta": {
+    "@mariozechner/pi-coding-agent": { "optional": true }
   },
   "devDependencies": {
     "@types/micromatch": "^4.0.10",
